Migration new architecture done :D
