--- masterAPP/app/src/main/java/vigi/care_provider/model/firebase/FirebaseConstants.java
@@ -1,7 +1,0 @@
-package vigi.care_provider.model.firebase;
-
-public final class FirebaseConstants {
-
-    public static final String IMAGE_PATH = "images/profiles";
-    public static final String JPG_EXTENSION = ".jpg";
-}
