--- masterAPP/app/src/main/java/vigi/care_provider/view/authentication/home/components/HomeAuthAnimatorListener.java
@@ -1,7 +1,0 @@
-package vigi.care_provider.view.authentication.home.components;
-
-import android.animation.AnimatorListenerAdapter;
-
-public class HomeAuthAnimatorListener extends AnimatorListenerAdapter {
-    //New further logic can be placed here
-}
