--- masterAPP/app/src/main/java/vigi/care_provider/view/utils/editText/EditTextUtils.java
@@ -1,14 +1,0 @@
-package vigi.care_provider.view.utils.editText;
-
-import android.widget.EditText;
-
-import java.util.Optional;
-
-public final class EditTextUtils {
-
-    public static String getTrimmedText(EditText editText) {
-        return Optional.ofNullable(editText.getText())
-                .map(CharSequence::toString)
-                .map(String::trim).get();
-    }
-}
