--- masterAPP/app/src/main/java/vigi/care_provider/view/vigi/activity/VigiActivity.java
@@ -1,9 +1,0 @@
-package vigi.care_provider.view.vigi.activity;
-
-public interface VigiActivity {
-
-    void setupUiComponents();
-
-    void setupClickListeners();
-
-}
