--- masterAPP/app/src/main/java/vigi/care_provider/view/vigi/activity/VigiLoginActivity.java
@@ -1,8 +1,0 @@
-package vigi.care_provider.view.vigi.activity;
-
-import vigi.care_provider.presenter.service.authentication.api.AuthenticationService;
-
-public interface VigiLoginActivity extends VigiActivity {
-
-    void performLogin(AuthenticationService authService, String email, String password);
-}
