+++ masterAPP/app/src/main/java/vigi/patient/model/firebase/FirebaseConstants.java
@@ -1,0 +1,7 @@
+package vigi.patient.model.firebase;
+
+public final class FirebaseConstants {
+
+    public static final String IMAGE_PATH = "images/profiles";
+    public static final String JPG_EXTENSION = ".jpg";
+}
