+++ masterAPP/app/src/main/java/vigi/patient/view/authentication/home/components/HomeAuthAnimatorListener.java
@@ -1,0 +1,7 @@
+package vigi.patient.view.authentication.home.components;
+
+import android.animation.AnimatorListenerAdapter;
+
+public class HomeAuthAnimatorListener extends AnimatorListenerAdapter {
+    //New further logic can be placed here
+}
