+++ masterAPP/app/src/main/java/vigi/patient/view/utils/editText/EditTextUtils.java
@@ -1,0 +1,14 @@
+package vigi.patient.view.utils.editText;
+
+import android.widget.EditText;
+
+import java.util.Optional;
+
+public final class EditTextUtils {
+
+    public static String getTrimmedText(EditText editText) {
+        return Optional.ofNullable(editText.getText())
+                .map(CharSequence::toString)
+                .map(String::trim).get();
+    }
+}
