+++ masterAPP/app/src/main/java/vigi/patient/view/vigi/activity/VigiActivity.java
@@ -1,0 +1,9 @@
+package vigi.patient.view.vigi.activity;
+
+public interface VigiActivity {
+
+    void setupUiComponents();
+
+    void setupClickListeners();
+
+}
