+++ masterAPP/app/src/main/java/vigi/patient/view/vigi/activity/VigiGenerateNewPasswordActivity.java
@@ -1,0 +1,8 @@
+package vigi.patient.view.vigi.activity;
+
+import vigi.patient.presenter.service.authentication.api.AuthenticationService;
+
+public interface VigiGenerateNewPasswordActivity extends VigiActivity {
+
+    void performGenerateNewPassword(AuthenticationService authService, String user);
+}
