+++ masterAPP/app/src/main/java/vigi/patient/view/vigi/activity/VigiLoginActivity.java
@@ -1,0 +1,8 @@
+package vigi.patient.view.vigi.activity;
+
+import vigi.patient.presenter.service.authentication.api.AuthenticationService;
+
+public interface VigiLoginActivity extends VigiActivity {
+
+    void performLogin(AuthenticationService authService, String email, String password);
+}
