+++ masterAPP/app/src/main/java/vigi/patient/view/vigi/activity/VigiRegisterActivity.java
@@ -1,0 +1,8 @@
+package vigi.patient.view.vigi.activity;
+
+import vigi.patient.presenter.service.authentication.api.AuthenticationService;
+
+public interface VigiRegisterActivity extends VigiActivity {
+
+    void performRegister(AuthenticationService authService, String email, String password);
+}
