+++ masterAPP/app/src/main/java/vigi/patient/view/vigi/web/VigiHtml.java
@@ -1,0 +1,18 @@
+package vigi.patient.view.vigi.web;
+
+import android.os.Build;
+import android.text.Html;
+import android.text.Spanned;
+
+public class VigiHtml {
+
+    @SuppressWarnings("deprecation")
+    public static Spanned fromHtml(String html) {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+            return Html.fromHtml(html, Html.FROM_HTML_MODE_LEGACY);
+        } else {
+            return Html.fromHtml(html);
+        }
+    }
+
+}
